Point Scootch Drive and Autonomous methods work mostly

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/DriveCommand.kt
@@ -5,7 +5,7 @@ package org.firstinspires.ftc.teamcode
 import kotlin.math.cos
 import kotlin.math.sin
 
-class DriveCommand(var xSpeed: Float, var ySpeed: Float, var rotationSpeed: Float) {
+class DriveCommand(var xSpeed: Double, var ySpeed: Double, var rotationSpeed: Double) {
 
     public fun rotate(radians: Float) {
         val x = xSpeed * cos(radians) - ySpeed * sin(radians)
